Began adding detailed target coverage report.
For every target, we'll report the mean coverage and coverage per base.
So far, I've gotten the data structure built in a way that has a consistent style as
the existing code (specifically how bed files are handled). And the stats are inside of
the capture_metrics object.

I've decided the output format will be
Key: target_name, mean_coverage bpcov,bpcov...
Key will likely be Cap_Target_Cov or something and target_name will be chr:start-end